Refactor checkPassword from PasswordService to IPasswordOperations
