+++ userdetails-gorm/src/main/groovy/au/org/ala/userdetails/gorm/GormPasswordOperations.groovy
@@ -3,6 +3,7 @@ package au.org.ala.userdetails.gorm
 import au.org.ala.auth.PasswordResetFailedException
 import au.org.ala.cas.encoding.BcryptPasswordEncoder
 import au.org.ala.cas.encoding.LegacyPasswordEncoder
+import au.org.ala.cas.encoding.PasswordEncoder
 import au.org.ala.userdetails.EmailService
 import au.org.ala.userdetails.IPasswordOperations
 import au.org.ala.users.UserRecord
@@ -13,6 +14,9 @@ class GormPasswordOperations implements IPasswordOperations {
 
     static final String BCRYPT_ENCODER_TYPE = 'bcrypt'
     static final String LEGACY_ENCODER_TYPE = 'legacy'
+
+    static final String STATUS_CURRENT = 'CURRENT'
+
 
     @Value('${password.encoder}')
     String passwordEncoderType = 'bcrypt'
@@ -76,4 +80,58 @@ class GormPasswordOperations implements IPasswordOperations {
         return "startPasswordReset"
     }
 
+    /**
+     * Check that a plain-text password matches a user's existing password.
+     * @param user The user.
+     * @param password The plain-text password to match.
+     * @return True if the password matches the existing password, otherwise false.
+     */
+    boolean checkUserPassword(UserRecord user, String password) {
+        assert user instanceof User
+        if (!password || password.size() < 1) {
+            throw new IllegalArgumentException("The password must not be empty.")
+        }
+        if (user == null) {
+            throw new IllegalArgumentException("Must provide the user to compare a password.")
+        }
+
+        def passwordType = getPasswordType()
+        def passwordStatus = STATUS_CURRENT
+        def existingPasswords = Password.findAllByUserAndTypeAndStatus(user, passwordType, passwordStatus)
+
+        def dateTimeNow = new Date().toTimestamp()
+
+        def matchingPassword = existingPasswords.find { item ->
+            comparePasswords(password, item.password) && (item.expiry == null || item.expiry > dateTimeNow)
+        }
+
+        return matchingPassword
+    }
+
+    /**
+     * Compare a plain-text password to an encoded password.
+     * @param plainPassword The plain-text password.
+     * @param hashedPassword The encoded password.
+     * @return True if the passwords match, otherwise false.
+     */
+    Boolean comparePasswords(String plainPassword, String hashedPassword) {
+        if (!plainPassword || plainPassword.length() < 1 || !hashedPassword || hashedPassword.length() < 1) {
+            throw new IllegalArgumentException("Must supply a plain text password and a hashed password to be compared.")
+        }
+
+        def encoder = getEncoder()
+        def encodedPassword = encoder.matches(plainPassword, hashedPassword)
+        return encodedPassword
+    }
+
+    private PasswordEncoder getEncoder() {
+        def encoder = passwordEncoderType.equalsIgnoreCase(BCRYPT_ENCODER_TYPE) ?
+                new BcryptPasswordEncoder(bcryptStrength) :
+                new LegacyPasswordEncoder(legacySalt, legacyAlgorithm, true)
+        return encoder
+    }
+
+    private getPasswordType() {
+        return passwordEncoderType.equalsIgnoreCase(BCRYPT_ENCODER_TYPE) ? BCRYPT_ENCODER_TYPE : LEGACY_ENCODER_TYPE
+    }
 }
